+++ doc/CHANGELOG.md
@@ -11,6 +11,7 @@
 ---
 ## HEAD <a name="head"></a>
 - `PS` initialized now with `000000` (see [ECO-028](ECO-028-ps_init.md))
+- `CPUERR` cleared with cpu reset (see [ECO-029](ECO-029-cpuerr_creset.md))
 
 <!-- --------------------------------------------------------------------- -->
 ---